Rename comp -> module
