+++ examples/resistors_and_nand.py
@@ -21,15 +21,15 @@ from faebryk.exporters.netlist.netlist import make_t2_netlist_from_t1
 # library imports
 from faebryk.library.core import Module
 from faebryk.library.kicad import KicadFootprint
-from faebryk.library.library.components import TI_CD4011BE, Resistor
 from faebryk.library.library.footprints import (
     SMDTwoPin,
     can_attach_to_footprint,
     can_attach_to_footprint_via_pinmap,
 )
 from faebryk.library.library.interfaces import Electrical, ElectricLogic, ElectricPower
+from faebryk.library.library.modules import TI_CD4011BE, Resistor
 from faebryk.library.library.parameters import Constant
-from faebryk.library.trait_impl.component import has_defined_type_description
+from faebryk.library.trait_impl.module import has_defined_type_description
 from faebryk.library.util import connect_interfaces_via_chain
 from faebryk.libs.experiments.buildutil import export_graph, export_netlist
 from faebryk.libs.logging import setup_basic_logging
