+++ src/faebryk/library/library/footprints.py
@@ -6,7 +6,7 @@ from abc import abstractmethod
 
 from faebryk.library.core import Footprint, FootprintTrait, ModuleInterface, ModuleTrait
 from faebryk.library.library.interfaces import Electrical
-from faebryk.library.trait_impl.component import has_defined_footprint
+from faebryk.library.trait_impl.module import has_defined_footprint
 from faebryk.libs.util import times
 
 logger = logging.getLogger(__name__)
